# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SCCC IT Department Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List entries (CIO sees all, others see own)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: weekOf
          in: query
          schema:
            type: string
            description: ISO date of the week start (Monday)
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Create a daily log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /entries/{id}:
    get:
      operationId: getEntry
      tags: [entries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
    put:
      operationId: updateEntry
      tags: [entries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
    delete:
      operationId: deleteEntry
      tags: [entries]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # WEEKLY REPORTS
...
components:
  schemas:
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userRole:
          type: string
        category:
          type: string
          enum: [helpdesk, network, security, general]
        title:
          type: string
        description:
          type: string
        accomplishments:
          type: string
        challenges:
          type: string
        supportNeeded:
          type: string
        ticketCount:
          type: integer
        weekOf:
          type: string
          description: ISO date string of week start
        entryDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        isSubmitted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        [
          id,
          userId,
          category,
          title,
          description,
          weekOf,
          isSubmitted,
          createdAt,
        ]
    CreateEntryBody:
      type: object
      properties:
        category:
          type: string
          enum: [helpdesk, network, security, general]
        title:
          type: string
        description:
          type: string
        accomplishments:
          type: string
        challenges:
          type: string
        supportNeeded:
          type: string
        ticketCount:
          type: integer
        weekOf:
          type: string
        entryDate:
          type: string
        tags:
          type: array
          items:
            type: string
        isSubmitted:
          type: boolean
      required: [category, title, description, weekOf]
...